chore: synchronize naftiko version to backstage
