+++ .github/workflows/synchronize-naftiko-version.yml
@@ -1,0 +1,141 @@
+name: Synchronize Naftiko Version
+
+on:
+  push:
+    branches:
+      - main
+    paths:
+      - 'pom.xml'
+  workflow_dispatch:
+
+
+jobs:
+  sync-framework-version:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+        with:
+          token: ${{ secrets.FRAMEWORK_CONTENTS_WRITE_PR_WRITE }}
+
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: '3.11'
+
+      - name: Set up Java
+        uses: actions/setup-java@v4
+        with:
+          distribution: 'temurin'
+          java-version: '17'
+
+      - name: Run version sync script
+        run: |
+          python3 src/main/resources/scripts/sync-naftiko-version.py
+
+      - name: Check for changes
+        id: git-check
+        run: |
+          if git diff --quiet; then
+            echo "changes=false" >> $GITHUB_OUTPUT
+          else
+            echo "changes=true" >> $GITHUB_OUTPUT
+          fi
+
+      - name: Create Pull Request
+        if: steps.git-check.outputs.changes == 'true'
+        uses: peter-evans/create-pull-request@v6
+        with:
+          token: ${{ secrets.FRAMEWORK_CONTENTS_WRITE_PR_WRITE }}
+          base: main
+          branch: "chore/sync-naftiko-version"
+          commit-message: "chore: sync Naftiko version from pom.xml [skip ci]"
+          title: "chore: sync Naftiko version"
+          body: |
+            ## No related issue
+
+            ---
+
+            ## What does this PR do?
+
+            Automatically synchronizes the Naftiko version across the repository:
+            - Updates `naftiko` version in YAML files
+            - Updates `const` in JSON schema
+            - Updates `$id` schema URL
+
+            Source of truth: `pom.xml`
+
+            ---
+
+            ## Tests
+
+            No additional tests required.
+            This change is limited to version synchronization and does not impact runtime behavior.
+
+            ---
+
+            ## Checklist
+
+            - [ ] CI is green (build, tests, schema validation, security scans)
+            - [ ] Rebased on latest `main`
+            - [ ] Small and focused — one concern per PR
+            - [ ] Commit messages follow [Conventional Commits](https://www.conventionalcommits.org/)
+          delete-branch: true
+
+  sync-backstage-version:
+    runs-on: ubuntu-latest
+    needs: sync-framework-version
+
+    steps:
+      - name: Checkout framework
+        uses: actions/checkout@v4
+        with:
+          path: framework
+
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: '3.11'
+
+      - name: Checkout backstage
+        uses: actions/checkout@v4
+        with:
+          repository: naftiko/backstage
+          token: ${{ secrets.BACKSTAGE_CONTENTS_WRITE_PR_WRITE }}
+          path: backstage
+
+      - name: Sync Naftiko version to backstage
+        id: sync
+        run: |
+          VERSION=$(python3 framework/src/main/resources/scripts/sync-naftiko-version-to-backstage.py \
+            --pom framework/pom.xml \
+            --target backstage/templates/skeletons/capabilities)
+          echo "value=${VERSION}" >> $GITHUB_OUTPUT
+
+      - name: Check for changes
+        id: git-check
+        working-directory: backstage
+        run: |
+          if git diff --quiet; then
+            echo "changes=false" >> $GITHUB_OUTPUT
+          else
+            echo "changes=true" >> $GITHUB_OUTPUT
+          fi
+
+      - name: Create Pull Request
+        if: steps.git-check.outputs.changes == 'true'
+        uses: peter-evans/create-pull-request@v6
+        with:
+          token: ${{ secrets.BACKSTAGE_CONTENTS_WRITE_PR_WRITE }}
+          path: backstage
+          base: main
+          branch: "chore/sync-naftiko-version"
+          commit-message: "chore: sync Naftiko version to ${{ steps.sync.outputs.value }} [skip ci]"
+          title: "chore: sync Naftiko version to ${{ steps.sync.outputs.value }}"
+          body: |
+            Automated sync from [framework@${{ github.sha }}](${{ github.server_url }}/${{ github.repository }}/commit/${{ github.sha }}).
+
+            Updated files:
+            - `templates/skeletons/capabilities/*.naftiko.yml`
+          delete-branch: true
